x/programs/cmd/simulator: move reusable simulator into its file
